+++ src/migrations/m190403_134317_fix_catchall_table.php
@@ -4,6 +4,7 @@ namespace venveo\redirect\migrations;
 
 use Craft;
 use craft\db\Migration;
+use craft\helpers\MigrationHelper;
 
 /**
  * m190403_134317_fix_catchall_table_and_rebrand migration.
@@ -17,7 +18,7 @@ class m190403_134317_fix_catchall_table extends Migration
     {
         // The erroneous table name...
         if ($this->db->tableExists('{{%dolphiq_redirects_catch_all_urls%}}')) {
-            $this->renameTable('{{%dolphiq_redirects_catch_all_urls%}}', '{{%dolphiq_redirects_catch_all_urls}}');
+            MigrationHelper::renameTable('{{%dolphiq_redirects_catch_all_urls%}}', '{{%dolphiq_redirects_catch_all_urls}}', $this);
         }
     }