+++ src/migrations/m190403_153859_add_redirect_type.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace venveo\redirect\migrations;
+
+use Craft;
+use craft\db\Migration;
+use craft\db\Table;
+
+/**
+ * m190403_153859_add_redirect_type migration.
+ */
+class m190403_153859_add_redirect_type extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function safeUp()
+    {
+        $this->addColumn('{{%dolphiq_redirects}}', 'type', $this->string('8')->null()->defaultValue('static')->notNull()->after('id'));
+        $this->createIndex($this->db->getIndexName('{{%dolphiq_redirects}}', 'type'), '{{%dolphiq_redirects}}', 'type');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown()
+    {
+        echo "m190403_153859_add_redirect_type cannot be reverted.\n";
+        return false;
+    }
+}