+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Forked for Venveo
 - Deleted unused code
 - Improved comments
+- Changed behavior of redirects to only fire when Craft encounters a 404 error
+- Optimized plugin code by bailing out early when possible to avoid additional calls
 
 ## 1.0.17 - 2018-07-04
 ### Fixed